move EsClient to own client.rs file
